feat: Implementa enrutamiento y refactoriza la estructura del proyecto
Esta mejora organiza la aplicación en páginas dedicadas y un módulo de utilidades.

- **Introduce enrutamiento** con `react-router-dom` para crear vistas separadas para "Panel de Presupuestos" y "Gestión de Personal".
- **Refactoriza el código** moviendo la lógica de presupuestos y el UI a un nuevo componente `BudgetDashboard.jsx`.
- **Extrae funciones auxiliares** a un nuevo archivo `utils.js` para mejorar la modularidad y la reutilización del código, solucionando el `ReferenceError` que ocurría.
- **Mejora la escalabilidad** y la mantenibilidad de la aplicación al separar las responsabilidades de los componentes.
